Separate interactive-or-not check from interactive input implementation
