+++ src/input/input_internal.h
@@ -6,7 +6,7 @@
 /*   By: amakinen <amakinen@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/18 15:23:45 by amakinen          #+#    #+#             */
-/*   Updated: 2025/06/18 15:24:30 by amakinen         ###   ########.fr       */
+/*   Updated: 2025/06/18 15:40:41 by amakinen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,5 +16,8 @@
 # include "status.h"
 
 t_status	input_notty_get_line(char **line_out, const char *prompt);
+t_status	input_tty_get_line(char **line_out, const char *prompt);
+void		input_tty_add_history(const char *line);
+void		input_tty_clear_history(void);
 
 #endif
