+++ src/input/input_tty.c
@@ -6,18 +6,16 @@
 /*   By: amakinen <amakinen@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/06 17:49:06 by amakinen          #+#    #+#             */
-/*   Updated: 2025/06/18 15:24:53 by amakinen         ###   ########.fr       */
+/*   Updated: 2025/06/18 15:41:53 by amakinen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "input.h"
 #include "input_internal.h"
 
 #include <errno.h>
 #include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <unistd.h>
 
 #include <readline/history.h>
 #include <readline/readline.h>
@@ -49,12 +47,10 @@ static int	input_rl_event_hook(void)
 	return (0);
 }
 
-t_status	input_get_line(char **line_out, const char *prompt)
+t_status	input_tty_get_line(char **line_out, const char *prompt)
 {
 	char	*line;
 
-	if (!isatty(STDIN_FILENO))
-		return (input_notty_get_line(line_out, prompt));
 	rl_startup_hook = input_rl_startup_hook;
 	rl_event_hook = input_rl_event_hook;
 	errno = 0;
@@ -71,13 +67,12 @@ t_status	input_get_line(char **line_out, const char *prompt)
 	return (S_OK);
 }
 
-void	input_add_history(const char *line)
+void	input_tty_add_history(const char *line)
 {
-	if (isatty(STDIN_FILENO))
-		add_history(line);
+	add_history(line);
 }
 
-void	input_clear_history(void)
+void	input_tty_clear_history(void)
 {
 	rl_clear_history();
 }
